remove redundant gc code
