+++ dbms/src/Storages/Page/ExternalPageCallbacks.cpp
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#pragma once
-
 #include <Storages/Page/ExternalPageCallbacks.h>
 
 namespace DB
